requirements.txt
torch==1.0.0
numpy==1.15.3
progressbar2==3.36.0
